app/pyproject.toml
[tool.poetry]
name = "cloud_run_cloud_sql_template"
version = "0.1.0"
description = ""
authors = ["Karn Wong <karn@karnwong.me>"]
readme = "README.md"
packages = [
    {include = "cloud_run_cloud_sql_template"}
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
